+++ hyper/install.sh
@@ -1,13 +1,12 @@
 #!/bin/bash -e
 
-if [[ ! -x "$(command -v hyper)" ]]
+if [[ "$(uname -s)" == "Darwin" ]] && [[ ! -x "/Applications/Hyper.app" ]]
 then
-  if [[ "$(uname -s)" == "Darwin" ]]
-  then
-    brew cask install hyper
-  else
-    wget --output-document=hyperterm.deb https://hyper-updates.now.sh/download/linux_deb
-    sudo dpkg -i hyperterm.deb
-    rm hyperterm.deb
-  fi
+  brew cask install hyper
+elif [[ "$(uname -s)" == "Linux" ]] && [[ ! -x "$(command -v hyper)" ]]
+then
+  # source
+  wget --output-document=hyperterm.deb https://hyper-updates.now.sh/download/linux_deb
+  sudo dpkg -i hyperterm.deb
+  rm hyperterm.deb
 fi
